Aggiunti mini-trigger

+++ servers/Customer/metodi/rimuoviCarrello.cpp
@@ -7,6 +7,11 @@ bool rimuoviCarrello(int idProdotto, int userID)
     Con2DB db(HOSTNAME, DB_PORT, USERNAME_CUST, PASSWORD_CUST, DB_NAME); // Connessione DB
     try
     {
+        sprintf(comando, "SELECT * FROM prodincart WHERE carrello = %d AND prodotto = %d", userID, idProdotto);
+        res = db.ExecSQLtuples(comando);
+        if (PQntuples(res) == 0) return false; // Se non ci sono prodotti nel carrello con quell'ID, impedisce l'operazione
+        PQclear(res);
+        // Rimuove il prodotto dal carrello
         sprintf(comando, "DELETE FROM prodincart WHERE carrello = %d AND prodotto = %d", userID, idProdotto);
         res = db.ExecSQLcmd(comando);
         PQclear(res);
@@ -15,6 +20,7 @@ bool rimuoviCarrello(int idProdotto, int userID)
     catch(...)
     {
         // Gestione errori database
+        PQclear(res);
         return false;
     }
 }
